refactor: extract tag management to separate component and store
