+++ src/App.tsx
@@ -1,114 +1,32 @@
 import { Button } from '@/components/ui/button';
 import { Card, CardContent } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { Textarea } from '@/components/ui/textarea';
 import { Calendar, Clock, MapPin, Send, Tag, Trash2 } from 'lucide-react';
-import React, { useState, useRef } from 'react';
+import { useState } from 'react';
+import { TagManager } from './feature/tag-manager/TagManager';
 import { Timeline } from './feature/timeline/Timeline';
+import { useDragAndDrop } from './hooks/use-drag-and-drop';
 import { useCharacterStore } from './store/character-store';
 import { useNotesStore } from './store/note-store';
 import { usePlaceStore } from './store/place-store';
 
-// タグの型定義
-interface Tag {
-  id: number;
-  type: 'character' | 'place' | 'time' | 'date';
-  text: string;
-  color?: string;
-}
-
 const DragDropMysteryApp = () => {
   // 状態管理
-  const { notes, addNote, deleteNote, addTagToNote, removeTagFromNote, setNotes } = useNotesStore();
-  const { characters, setCharacters, getCharacterColor } = useCharacterStore();
-  const { places, setPlaces, getPlaceColor } = usePlaceStore();
-  const [quickTimes, setQuickTimes] = useState<string[]>([]);
-  const [quickDates, setQuickDates] = useState<string[]>([]);
+  const { notes, addNote, deleteNote, removeTagFromNote, setNotes } = useNotesStore();
+  const { getCharacterColor } = useCharacterStore();
+  const { getPlaceColor } = usePlaceStore();
   const [newNote, setNewNote] = useState('');
-  const [newCharacterText, setNewCharacterText] = useState('');
-  const [newPlaceText, setNewPlaceText] = useState('');
-  const [newTimeText, setNewTimeText] = useState('');
-  const [newDateText, setNewDateText] = useState('');
-  const [newTagColor, setNewTagColor] = useState('#000000');
   const [rightPanelMode, setRightPanelMode] = useState<'tags' | 'timeline'>('tags');
   const [editingNote, setEditingNote] = useState<number | null>(null);
   const [editNoteContent, setEditNoteContent] = useState('');
-  const [editingTagId, setEditingTagId] = useState<number | null>(null);
-  const [editingTagType, setEditingTagType] = useState<'character' | 'place' | 'time' | 'date' | null>(null);
-  const [editTagText, setEditTagText] = useState('');
-  const [editTagColor, setEditTagColor] = useState('');
 
-  const dragItem = useRef<Tag | null>(null);
-  const dragOverItem = useRef<number | null>(null);
+  const { handleDragOver, handleDragLeave, handleDrop } = useDragAndDrop();
 
   // メモを追加
   const handleAddNote = () => {
     addNote(newNote);
     setNewNote('');
-  };
-
-  // タグをドラッグ開始時のハンドラ
-  const handleDragStart = (e: React.DragEvent<HTMLDivElement>, tag: Tag) => {
-    // タグの情報をドラッグデータに設定
-    e.dataTransfer.setData('application/json', JSON.stringify(tag));
-    dragItem.current = tag;
-
-    // ドラッグ中の見た目の設定
-    if (e.currentTarget.classList) {
-      setTimeout(() => {
-        e.currentTarget.classList.add('opacity-50');
-      }, 0);
-    }
-  };
-
-  // ドラッグ終了時のハンドラー
-  const handleDragEnd = (e: React.DragEvent<HTMLDivElement>) => {
-    if (e.currentTarget.classList) {
-      e.currentTarget.classList.remove('opacity-50');
-    }
-    dragItem.current = null;
-    dragOverItem.current = null;
-  };
-
-  // ドラッグ中にメモの上に来た時のハンドラー
-  const handleDragOver = (e: React.DragEvent<HTMLDivElement>, noteId: number) => {
-    e.preventDefault();
-    dragOverItem.current = noteId;
-
-    // ドロップ可能エリアの視覚的なフィードバック
-    if (e.currentTarget.classList) {
-      e.currentTarget.classList.add('border-primary');
-    }
-  };
-
-  // ドラッグがメモの上から離れた時のハンドラー
-  const handleDragLeave = (e: React.DragEvent<HTMLDivElement>) => {
-    if (e.currentTarget.classList) {
-      e.currentTarget.classList.remove('border-primary');
-    }
-  };
-
-  // メモにタグをドロップした時のハンドラー
-  const handleDrop = (e: React.DragEvent<HTMLDivElement>, noteId: number) => {
-    e.preventDefault();
-
-    // ドロップされたタグのデータを取得
-    const tagData = JSON.parse(e.dataTransfer.getData('application/json')) as Tag;
-
-    // ドロップされたタグの種類に応じて処理
-    if (tagData) {
-      addTagToNote(noteId, tagData.type, tagData.text);
-    }
-
-    // ドラッグ関連の参照をクリア
-    dragItem.current = null;
-    dragOverItem.current = null;
-
-    // ドロップターゲットのスタイルをリセット
-    if (e.currentTarget) {
-      e.currentTarget.classList.remove('bg-gray-100');
-    }
   };
 
   // タグをノートから削除
@@ -128,662 +46,6 @@ const DragDropMysteryApp = () => {
       e.preventDefault();
       handleAddNote();
     }
-  };
-
-  // タグを編集
-  const startEditingTag = (id: number, type: 'character' | 'place' | 'time' | 'date', text: string, color?: string) => {
-    setEditingTagId(id);
-    setEditingTagType(type);
-    setEditTagText(text);
-    if (color) setEditTagColor(color);
-  };
-
-  const saveTagEdit = () => {
-    if (!editingTagType || editingTagId === null || editTagText.trim() === '') return;
-
-    // 古い名前を保持しておく
-    let oldName = '';
-
-    switch (editingTagType) {
-      case 'character':
-        // 変更前の名前を取得
-        oldName = characters.find((char) => char.id === editingTagId)?.name || '';
-
-        // キャラクター情報を更新
-        setCharacters(
-          characters.map((char) =>
-            char.id === editingTagId ? { ...char, name: editTagText, color: editTagColor } : char,
-          ),
-        );
-
-        // 関連するメモも更新
-        if (oldName) {
-          setNotes(notes.map((note) => (note.character === oldName ? { ...note, character: editTagText } : note)));
-        }
-        break;
-      case 'place':
-        // 変更前の名前を取得
-        oldName = places.find((place) => place.id === editingTagId)?.name || '';
-
-        // 場所情報を更新
-        setPlaces(
-          places.map((place) =>
-            place.id === editingTagId ? { ...place, name: editTagText, color: editTagColor } : place,
-          ),
-        );
-
-        // 関連するメモも更新
-        if (oldName) {
-          setNotes(notes.map((note) => (note.place === oldName ? { ...note, place: editTagText } : note)));
-        }
-        break;
-      case 'time':
-        // 変更前の時間を取得
-        oldName = quickTimes[editingTagId];
-
-        // 時間情報を更新
-        setQuickTimes(quickTimes.map((time, index) => (index === editingTagId ? editTagText : time)));
-
-        // 関連するメモも更新
-        if (oldName) {
-          setNotes(notes.map((note) => (note.time === oldName ? { ...note, time: editTagText } : note)));
-        }
-        break;
-      case 'date':
-        // 変更前の日付を取得
-        oldName = quickDates[editingTagId];
-
-        // 日付情報を更新
-        setQuickDates(quickDates.map((date, index) => (index === editingTagId ? editTagText : date)));
-
-        // 関連するメモも更新
-        if (oldName) {
-          setNotes(notes.map((note) => (note.date === oldName ? { ...note, date: editTagText } : note)));
-        }
-        break;
-    }
-
-    // 編集状態をリセット
-    setEditingTagId(null);
-    setEditingTagType(null);
-    setEditTagText('');
-  };
-
-  // タグを削除
-  const deleteTag = (id: number, type: 'character' | 'place' | 'time' | 'date') => {
-    switch (type) {
-      case 'character':
-        setCharacters(characters.filter((char) => char.id !== id));
-        // キャラクターを使用しているノートからタグを削除
-        setNotes(
-          notes.map((note) => {
-            const charName = characters.find((c) => c.id === id)?.name;
-            if (note.character === charName) {
-              return { ...note, character: null };
-            }
-            return note;
-          }),
-        );
-        break;
-      case 'place':
-        setPlaces(places.filter((place) => place.id !== id));
-        // 場所を使用しているノートからタグを削除
-        setNotes(
-          notes.map((note) => {
-            const placeName = places.find((p) => p.id === id)?.name;
-            if (note.place === placeName) {
-              return { ...note, place: null };
-            }
-            return note;
-          }),
-        );
-        break;
-      case 'time':
-        const timeValue = quickTimes[id];
-        setQuickTimes(quickTimes.filter((_, index) => index !== id));
-        // 時間を使用しているノートからタグを削除
-        setNotes(
-          notes.map((note) => {
-            if (note.time === timeValue) {
-              return { ...note, time: null };
-            }
-            return note;
-          }),
-        );
-        break;
-      case 'date':
-        const dateValue = quickDates[id];
-        setQuickDates(quickDates.filter((_, index) => index !== id));
-        // 日付を使用しているノートからタグを削除
-        setNotes(
-          notes.map((note) => {
-            if (note.date === dateValue) {
-              return { ...note, date: null };
-            }
-            return note;
-          }),
-        );
-        break;
-    }
-  };
-
-  // タグパレットのレンダリング
-  const renderTagPalette = () => {
-    return (
-      <div className="flex flex-col h-full">
-        <ScrollArea className="flex-1 max-h-[calc(100vh-12rem)]">
-          {/* キャラクタータグ */}
-          <div className="mb-4">
-            <h3 className="text-sm font-semibold mb-2">キャラクター</h3>
-            <div className="flex items-center gap-2 mb-2">
-              <Input
-                value={newCharacterText}
-                onChange={(e) => setNewCharacterText(e.target.value)}
-                placeholder="新しいキャラクターを追加"
-                className="flex-1 text-sm"
-              />
-              <Input
-                type="color"
-                value={newTagColor}
-                onChange={(e) => setNewTagColor(e.target.value)}
-                className="w-10"
-              />
-              <Button
-                variant="outline"
-                size="sm"
-                onClick={() => {
-                  if (newCharacterText.trim() === '') return;
-                  const newChar = {
-                    id: Date.now(),
-                    name: newCharacterText,
-                    color: newTagColor,
-                  };
-                  setCharacters([...characters, newChar]);
-                  setNewCharacterText('');
-                }}
-              >
-                追加
-              </Button>
-            </div>
-            <div className="grid grid-cols-3 gap-1.5">
-              {characters.map((char) => (
-                <div
-                  key={char.id}
-                  className="relative bg-card border rounded-md px-2 py-1 flex items-center gap-1 text-xs group"
-                  style={{ borderLeftColor: char.color, borderLeftWidth: '4px' }}
-                >
-                  {editingTagId === char.id && editingTagType === 'character' ? (
-                    <div className="flex items-center w-full gap-1">
-                      <Input
-                        value={editTagText}
-                        onChange={(e) => setEditTagText(e.target.value)}
-                        className="h-5 text-xs px-1 py-0 w-full"
-                        autoFocus
-                        onKeyDown={(e) => {
-                          if (e.key === 'Enter') saveTagEdit();
-                          if (e.key === 'Escape') {
-                            setEditingTagId(null);
-                            setEditingTagType(null);
-                          }
-                        }}
-                      />
-                      <Input
-                        type="color"
-                        value={editTagColor}
-                        onChange={(e) => setEditTagColor(e.target.value)}
-                        className="w-5 h-5 p-0"
-                        onClick={(e) => e.stopPropagation()}
-                      />
-                      <div className="flex ml-1">
-                        <button
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            saveTagEdit();
-                          }}
-                          className="text-green-500 hover:text-green-700 p-0.5"
-                        >
-                          <svg
-                            width="12"
-                            height="12"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                          >
-                            <path d="M9 16.17L4.83 12l-1.42 1.41L9 19 21 7l-1.41-1.41L9 16.17z" fill="currentColor" />
-                          </svg>
-                        </button>
-                        <button
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            setEditingTagId(null);
-                            setEditingTagType(null);
-                          }}
-                          className="text-red-500 hover:text-red-700 p-0.5"
-                        >
-                          <svg
-                            width="12"
-                            height="12"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                          >
-                            <path
-                              d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12 19 6.41z"
-                              fill="currentColor"
-                            />
-                          </svg>
-                        </button>
-                      </div>
-                    </div>
-                  ) : (
-                    <>
-                      <div
-                        draggable
-                        onDragStart={(e) =>
-                          handleDragStart(e, { id: char.id, type: 'character', text: char.name, color: char.color })
-                        }
-                        onDragEnd={handleDragEnd}
-                        className="flex items-center gap-1 cursor-move flex-1"
-                      >
-                        <div className="w-2 h-2 rounded-full" style={{ backgroundColor: char.color }}></div>
-                        <span className="truncate">{char.name}</span>
-                      </div>
-                      <div className="opacity-0 group-hover:opacity-100 flex">
-                        <button
-                          onClick={() => startEditingTag(char.id, 'character', char.name, char.color)}
-                          className="text-gray-500 hover:text-gray-700 p-0.5"
-                        >
-                          <svg
-                            width="12"
-                            height="12"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                          >
-                            <path
-                              d="M16.474 5.408L18.592 7.526L6 20.118V22H8.118L20.71 9.408L22.828 11.526L23.535 10.819L13.181 0.465L12.474 1.172L14.592 3.29L16.474 5.408ZM13.185 7.112L13.644 7.571L15.354 9.281L18.829 5.807L16.647 3.625L13.185 7.112Z"
-                              fill="currentColor"
-                            />
-                          </svg>
-                        </button>
-                        <button
-                          onClick={() => deleteTag(char.id, 'character')}
-                          className="text-red-500 hover:text-red-700 p-0.5"
-                        >
-                          <svg
-                            width="12"
-                            height="12"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                          >
-                            <path
-                              d="M20 6H16V5C16 3.9 15.1 3 14 3H10C8.9 3 8 3.9 8 5V6H4C2.9 6 2 6.9 2 8V9C2 9.6 2.4 10 3 10H4V20C4 21.1 4.9 22 6 22H18C19.1 22 20 21.1 20 20V10H21C21.6 10 22 9.6 22 9V8C22 6.9 21.1 6 20 6ZM10 5H14V6H10V5ZM18 20H6V10H18V20Z"
-                              fill="currentColor"
-                            />
-                          </svg>
-                        </button>
-                      </div>
-                    </>
-                  )}
-                </div>
-              ))}
-            </div>
-          </div>
-
-          {/* 場所タグ */}
-          <div className="mb-4">
-            <h3 className="text-sm font-semibold mb-2">場所</h3>
-            <div className="flex items-center gap-2 mb-2">
-              <Input
-                value={newPlaceText}
-                onChange={(e) => setNewPlaceText(e.target.value)}
-                placeholder="新しい場所を追加"
-                className="flex-1 text-sm"
-              />
-              <Input
-                type="color"
-                value={newTagColor}
-                onChange={(e) => setNewTagColor(e.target.value)}
-                className="w-10"
-              />
-              <Button
-                variant="outline"
-                size="sm"
-                onClick={() => {
-                  if (newPlaceText.trim() === '') return;
-                  const newPlace = {
-                    id: Date.now(),
-                    name: newPlaceText,
-                    color: newTagColor,
-                  };
-                  setPlaces([...places, newPlace]);
-                  setNewPlaceText('');
-                }}
-              >
-                追加
-              </Button>
-            </div>
-            <div className="grid grid-cols-3 gap-1.5">
-              {places.map((place) => (
-                <div
-                  key={place.id}
-                  className="relative bg-card border rounded-md px-2 py-1 flex items-center gap-1 text-xs group"
-                  style={{ borderLeftColor: place.color, borderLeftWidth: '4px' }}
-                >
-                  {editingTagId === place.id && editingTagType === 'place' ? (
-                    <div className="flex items-center w-full gap-1">
-                      <Input
-                        value={editTagText}
-                        onChange={(e) => setEditTagText(e.target.value)}
-                        className="h-5 text-xs px-1 py-0 w-full"
-                        autoFocus
-                        onKeyDown={(e) => {
-                          if (e.key === 'Enter') saveTagEdit();
-                          if (e.key === 'Escape') {
-                            setEditingTagId(null);
-                            setEditingTagType(null);
-                          }
-                        }}
-                      />
-                      <Input
-                        type="color"
-                        value={editTagColor}
-                        onChange={(e) => setEditTagColor(e.target.value)}
-                        className="w-5 h-5 p-0"
-                        onClick={(e) => e.stopPropagation()}
-                      />
-                      <div className="flex ml-1">
-                        <button
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            saveTagEdit();
-                          }}
-                          className="text-green-500 hover:text-green-700 p-0.5"
-                        >
-                          <svg
-                            width="12"
-                            height="12"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                          >
-                            <path d="M9 16.17L4.83 12l-1.42 1.41L9 19 21 7l-1.41-1.41L9 16.17z" fill="currentColor" />
-                          </svg>
-                        </button>
-                        <button
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            setEditingTagId(null);
-                            setEditingTagType(null);
-                          }}
-                          className="text-red-500 hover:text-red-700 p-0.5"
-                        >
-                          <svg
-                            width="12"
-                            height="12"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                          >
-                            <path
-                              d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12 19 6.41z"
-                              fill="currentColor"
-                            />
-                          </svg>
-                        </button>
-                      </div>
-                    </div>
-                  ) : (
-                    <>
-                      <div
-                        draggable
-                        onDragStart={(e) =>
-                          handleDragStart(e, { id: place.id, type: 'place', text: place.name, color: place.color })
-                        }
-                        onDragEnd={handleDragEnd}
-                        className="flex items-center gap-1 cursor-move flex-1"
-                      >
-                        <MapPin className="h-2 w-2" style={{ color: place.color }} />
-                        <span className="truncate">{place.name}</span>
-                      </div>
-                      <div className="opacity-0 group-hover:opacity-100 flex">
-                        <button
-                          onClick={() => startEditingTag(place.id, 'place', place.name, place.color)}
-                          className="text-gray-500 hover:text-gray-700 p-0.5"
-                        >
-                          <svg
-                            width="12"
-                            height="12"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                          >
-                            <path
-                              d="M16.474 5.408L18.592 7.526L6 20.118V22H8.118L20.71 9.408L22.828 11.526L23.535 10.819L13.181 0.465L12.474 1.172L14.592 3.29L16.474 5.408ZM13.185 7.112L13.644 7.571L15.354 9.281L18.829 5.807L16.647 3.625L13.185 7.112Z"
-                              fill="currentColor"
-                            />
-                          </svg>
-                        </button>
-                        <button
-                          onClick={() => deleteTag(place.id, 'place')}
-                          className="text-red-500 hover:text-red-700 p-0.5"
-                        >
-                          <svg
-                            width="12"
-                            height="12"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                          >
-                            <path
-                              d="M20 6H16V5C16 3.9 15.1 3 14 3H10C8.9 3 8 3.9 8 5V6H4C2.9 6 2 6.9 2 8V9C2 9.6 2.4 10 3 10H4V20C4 21.1 4.9 22 6 22H18C19.1 22 20 21.1 20 20V10H21C21.6 10 22 9.6 22 9V8C22 6.9 21.1 6 20 6ZM10 5H14V6H10V5ZM18 20H6V10H18V20Z"
-                              fill="currentColor"
-                            />
-                          </svg>
-                        </button>
-                      </div>
-                    </>
-                  )}
-                </div>
-              ))}
-            </div>
-          </div>
-
-          {/* 時間タグ */}
-          <div className="mb-4">
-            <h3 className="text-sm font-semibold mb-2">時間</h3>
-            <div className="flex items-center gap-2 mb-2">
-              <Input
-                value={newTimeText}
-                onChange={(e) => setNewTimeText(e.target.value)}
-                placeholder="新しい時間を追加"
-                className="flex-1 text-sm"
-              />
-              <Button
-                variant="outline"
-                size="sm"
-                onClick={() => {
-                  if (newTimeText.trim() === '' || quickTimes.includes(newTimeText)) return;
-                  setQuickTimes([...quickTimes, newTimeText]);
-                  setNewTimeText('');
-                }}
-              >
-                追加
-              </Button>
-            </div>
-            <div className="grid grid-cols-4 gap-1.5">
-              {quickTimes.map((time, index) => (
-                <div
-                  key={index}
-                  className="relative bg-card border rounded-md px-2 py-1 flex items-center gap-1 text-xs group"
-                >
-                  {editingTagId === index && editingTagType === 'time' ? (
-                    <Input
-                      value={editTagText}
-                      onChange={(e) => setEditTagText(e.target.value)}
-                      className="h-5 text-xs px-1 py-0 w-full"
-                      autoFocus
-                      onKeyDown={(e) => {
-                        if (e.key === 'Enter') saveTagEdit();
-                        if (e.key === 'Escape') {
-                          setEditingTagId(null);
-                          setEditingTagType(null);
-                        }
-                      }}
-                    />
-                  ) : (
-                    <>
-                      <div
-                        draggable
-                        onDragStart={(e) => handleDragStart(e, { id: index, type: 'time', text: time })}
-                        onDragEnd={handleDragEnd}
-                        className="flex items-center gap-1 cursor-move flex-1"
-                      >
-                        <Clock className="h-2 w-2 text-gray-500" />
-                        <span className="truncate">{time}</span>
-                      </div>
-                      <div className="opacity-0 group-hover:opacity-100 flex">
-                        <button
-                          onClick={() => startEditingTag(index, 'time', time)}
-                          className="text-gray-500 hover:text-gray-700 p-0.5"
-                        >
-                          <svg
-                            width="12"
-                            height="12"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                          >
-                            <path
-                              d="M16.474 5.408L18.592 7.526L6 20.118V22H8.118L20.71 9.408L22.828 11.526L23.535 10.819L13.181 0.465L12.474 1.172L14.592 3.29L16.474 5.408ZM13.185 7.112L13.644 7.571L15.354 9.281L18.829 5.807L16.647 3.625L13.185 7.112Z"
-                              fill="currentColor"
-                            />
-                          </svg>
-                        </button>
-                        <button
-                          onClick={() => deleteTag(index, 'time')}
-                          className="text-red-500 hover:text-red-700 p-0.5"
-                        >
-                          <svg
-                            width="12"
-                            height="12"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                          >
-                            <path
-                              d="M20 6H16V5C16 3.9 15.1 3 14 3H10C8.9 3 8 3.9 8 5V6H4C2.9 6 2 6.9 2 8V9C2 9.6 2.4 10 3 10H4V20C4 21.1 4.9 22 6 22H18C19.1 22 20 21.1 20 20V10H21C21.6 10 22 9.6 22 9V8C22 6.9 21.1 6 20 6ZM10 5H14V6H10V5ZM18 20H6V10H18V20Z"
-                              fill="currentColor"
-                            />
-                          </svg>
-                        </button>
-                      </div>
-                    </>
-                  )}
-                </div>
-              ))}
-            </div>
-          </div>
-
-          {/* 日付タグ */}
-          <div className="mb-4">
-            <h3 className="text-sm font-semibold mb-2">日付</h3>
-            <div className="flex items-center gap-2 mb-2">
-              <Input
-                value={newDateText}
-                onChange={(e) => setNewDateText(e.target.value)}
-                placeholder="新しい日付を追加 (例: 1日目)"
-                className="flex-1 text-sm"
-              />
-              <Button
-                variant="outline"
-                size="sm"
-                onClick={() => {
-                  if (newDateText.trim() === '' || quickDates.includes(newDateText)) return;
-                  setQuickDates([...quickDates, newDateText]);
-                  setNewDateText('');
-                }}
-              >
-                追加
-              </Button>
-            </div>
-            <div className="grid grid-cols-4 gap-1.5">
-              {quickDates.map((date, index) => (
-                <div
-                  key={index}
-                  className="relative bg-card border rounded-md px-2 py-1 flex items-center gap-1 text-xs group"
-                >
-                  {editingTagId === index && editingTagType === 'date' ? (
-                    <Input
-                      value={editTagText}
-                      onChange={(e) => setEditTagText(e.target.value)}
-                      className="h-5 text-xs px-1 py-0 w-full"
-                      autoFocus
-                      onKeyDown={(e) => {
-                        if (e.key === 'Enter') saveTagEdit();
-                        if (e.key === 'Escape') {
-                          setEditingTagId(null);
-                          setEditingTagType(null);
-                        }
-                      }}
-                    />
-                  ) : (
-                    <>
-                      <div
-                        draggable
-                        onDragStart={(e) => handleDragStart(e, { id: index, type: 'date', text: date })}
-                        onDragEnd={handleDragEnd}
-                        className="flex items-center gap-1 cursor-move flex-1"
-                      >
-                        <Calendar className="h-2 w-2 text-gray-500" />
-                        <span className="truncate">{date}</span>
-                      </div>
-                      <div className="opacity-0 group-hover:opacity-100 flex">
-                        <button
-                          onClick={() => startEditingTag(index, 'date', date)}
-                          className="text-gray-500 hover:text-gray-700 p-0.5"
-                        >
-                          <svg
-                            width="12"
-                            height="12"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                          >
-                            <path
-                              d="M16.474 5.408L18.592 7.526L6 20.118V22H8.118L20.71 9.408L22.828 11.526L23.535 10.819L13.181 0.465L12.474 1.172L14.592 3.29L16.474 5.408ZM13.185 7.112L13.644 7.571L15.354 9.281L18.829 5.807L16.647 3.625L13.185 7.112Z"
-                              fill="currentColor"
-                            />
-                          </svg>
-                        </button>
-                        <button
-                          onClick={() => deleteTag(index, 'date')}
-                          className="text-red-500 hover:text-red-700 p-0.5"
-                        >
-                          <svg
-                            width="12"
-                            height="12"
-                            viewBox="0 0 24 24"
-                            fill="none"
-                            xmlns="http://www.w3.org/2000/svg"
-                          >
-                            <path
-                              d="M20 6H16V5C16 3.9 15.1 3 14 3H10C8.9 3 8 3.9 8 5V6H4C2.9 6 2 6.9 2 8V9C2 9.6 2.4 10 3 10H4V20C4 21.1 4.9 22 6 22H18C19.1 22 20 21.1 20 20V10H21C21.6 10 22 9.6 22 9V8C22 6.9 21.1 6 20 6ZM10 5H14V6H10V5ZM18 20H6V10H18V20Z"
-                              fill="currentColor"
-                            />
-                          </svg>
-                        </button>
-                      </div>
-                    </>
-                  )}
-                </div>
-              ))}
-            </div>
-          </div>
-        </ScrollArea>
-      </div>
-    );
   };
 
   return (
@@ -988,7 +250,7 @@ const DragDropMysteryApp = () => {
           </div>
         </div>
 
-        {rightPanelMode === 'tags' ? renderTagPalette() : <Timeline />}
+        {rightPanelMode === 'tags' ? <TagManager /> : <Timeline />}
       </div>
     </div>
   );
