+++ src/store/time-store.ts
@@ -1,0 +1,11 @@
+import { create } from 'zustand';
+
+interface TimeState {
+  times: string[];
+  setTimes: (times: string[]) => void;
+}
+
+export const useTimeStore = create<TimeState>((set) => ({
+  times: [],
+  setTimes: (times: string[]) => set({ times }),
+}));
